fix(automate-linking): missing subfields in response
- utilize common logic for string-to-subfields parsing

Closes: MODQM-358
