+++ src/main/java/org/folio/qm/util/MarcUtils.java
@@ -1,16 +1,29 @@
 package org.folio.qm.util;
 
+import static java.util.Objects.requireNonNull;
+import static org.apache.commons.lang3.StringUtils.EMPTY;
+import static org.folio.qm.converter.elements.Constants.CONCAT_CONDITION_PATTERN;
+import static org.folio.qm.converter.elements.Constants.SPLIT_PATTERN;
+import static org.folio.qm.converter.elements.Constants.TOKEN_MIN_LENGTH;
+
 import com.google.common.collect.BiMap;
 import com.google.common.collect.ImmutableBiMap;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Optional;
+import java.util.function.Function;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 import org.apache.commons.lang3.StringUtils;
 import org.folio.qm.domain.dto.BaseMarcRecord;
 import org.folio.qm.domain.dto.FieldItem;
 import org.folio.qm.domain.dto.MarcFormat;
 import org.folio.qm.domain.dto.ParsedRecordDto;
+import org.marc4j.marc.Subfield;
 
 public final class MarcUtils {
   public static final BiMap<ParsedRecordDto.RecordTypeEnum, MarcFormat> TYPE_MAP = ImmutableBiMap.of(
@@ -84,4 +97,28 @@ public final class MarcUtils {
   public static boolean isValidUuid(String id) {
     return UUID_REGEX.matcher(id).matches();
   }
+
+  public static List<Subfield> extractSubfields(FieldItem field, Function<String, Subfield> subfieldFunction) {
+    var tokens = Arrays.stream(SPLIT_PATTERN.split(field.getContent().toString()))
+      .collect(Collectors.toCollection(LinkedList::new));
+
+    List<Subfield> subfields = new ArrayList<>();
+    while (!tokens.isEmpty()) {
+      String token = tokens.pop();
+      String subfieldString = token.concat(checkNextToken(tokens));
+      if (subfieldString.length() < TOKEN_MIN_LENGTH) {
+        throw new IllegalArgumentException("Subfield length");
+      }
+      subfields.add(subfieldFunction.apply(subfieldString));
+    }
+
+    return subfields;
+  }
+
+  private static String checkNextToken(LinkedList<String> tokens) {
+    return !tokens.isEmpty() && CONCAT_CONDITION_PATTERN.matcher(tokens.peek()).matches()
+           ? requireNonNull(tokens.poll()).concat(checkNextToken(tokens))
+           : EMPTY;
+  }
+
 }
